Fix passing arguments to a generator function
Before this commit the part of the generator function that handles
declaring arguments was moved to the callback. This essentially made the
arguments unusable because callbacks have their own arguments. In order
to make it work properly we need to set variables on the scope before we
switch to the callback. This is achieved by adding a dummy function as a
"marker" that we can use to figure out which part to keep in the
gnerator function