Fix setting properties on an object with a prototype
The set_property function was using get_property to fetch an *existing*
property. This means, it could also fetch an existing property from one
of the parent prototypes, and then it would update its value. In order
to fix it, we have to use get_own_property that will only search through
object properties.